requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
deep-sort-realtime==1.3.2
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
idna==3.10
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycairo==1.28.0
pydantic==2.11.4
pydantic_core==2.33.2
PyGObject==3.52.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reactivex==4.0.4
requests==2.32.3
scipy==1.15.3
seaborn==0.13.2
setuptools==80.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.137
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
